==> 08_map_price_electricity.Rmd <==
---
title: "Analysis of ENCEVI 2018"
author: '[Mauricio Hernandez](http://mauricioh2.com)'
date: "`r Sys.Date()`"
subtitle: Price of Electricity Map
bibliography: energy_mex.bib
---


```{r, include=FALSE}
#Remove all objects from current workspace and call garbage collector
rm(list=ls())
gc()

source("./script/general_setup_functions.R")
```


```{r knitr_init, include=FALSE}
#options(max.print = "75")
opts_chunk$set(echo = TRUE,
	             cache = FALSE,
               prompt = FALSE,
               tidy = FALSE,
               comment = NA,
               message = FALSE,
               warning = FALSE,
               fig.path = 'figures/')
```


```{r setup, include=FALSE}
#Generates the css needed by summarytools
st_css()

#devtools::install_github("ropensci/plotly")
knitr::opts_chunk$set(fig.width = 8, collapse = TRUE)

knitr::knit_hooks$set(output = function(x, options){
  if(!is.null(options$max_height)){
    paste('<pre style = "max-height:',
          options$max_height, 
          ';float: left; width: 910px; overflow-y: auto;">',
          x,
          "</pre>",
          sep = "")
  } else{
    x
  }
})
```


```{r}
#Libraries to build map
if (!require("pacman")) install.packages("pacman")
pacman::p_load(party, magrittr, ggplot2,
               knitr, foreign, mxmaps, openxlsx, 
               viridis, scales, geojsonio, rgdal, 
               rgeos, maptools, ggrepel, sf)
```

**Loading datasets**
</br>This dataset was obtained with the "05_bill_power_calculation.Rmd" file.
```{r}
df_clim_reg<- read.csv("./input/state_climatic_regions.csv", 
                          header = TRUE, na.strings="")

df_bill_cons <- read.csv("./output/df_bill_wconsump.csv", 
                           header = TRUE, na.strings="")

df_bill_cons <- df_bill_cons %>%
  rename(state.name = state.f)

head(df_bill_cons)
```

```{r}
# create state code variable as a char
df_clim_reg$state_code <- as.character(df_clim_reg$state.code)
df_clim_reg$state_code[df_clim_reg$state.code < 10] <- paste0('0',df_clim_reg$state.code)
head(df_clim_reg)
```

```{r}
MXN_TO_USD = 18.70

#Price paid after fees and taxes
df_bill_cons$price.kWh.med1 <- df_bill_cons$bill.amount1 / df_bill_cons$cons.kWh.med1
df_bill_cons$price.mWh.usd <- (df_bill_cons$price.kWh.med1 * 1000) / MXN_TO_USD

#Price paid before fees and taxes
df_bill_cons$price.kWh.notax.med1 <- df_bill_cons$bill.avg1.notax / df_bill_cons$cons.kWh.med1
df_bill_cons$price.mWh.notax.usd <- (df_bill_cons$price.kWh.notax.med1 * 1000) / MXN_TO_USD
```

```{r}
options(survey.lonely.psu = "adjust")

#df_bill_cons %>% drop_na(is.summer1)

# Create survey design
dstrata <- df_bill_cons %>%
  as_survey_design(id = upm, strata = est_dis, weights = factor_sem)
```


**Get mean cost per MWh After Taxes**
```{r}

tbl_mean_cost_state <- dstrata %>% 
  group_by(state.name, is.summer1)  %>%
  summarize(mean_cost_mwh = survey_mean(price.mWh.usd, na.rm= TRUE, vartype = c("ci"))) 

tbl_mean_cost_state <- na.omit(tbl_mean_cost_state) 
colnames(tbl_mean_cost_state)[which(names(tbl_mean_cost_state) == "region.f")] <- "Region"
# renaming all the column names of data frame

#State names that are different in both dataframe
tbl_mean_cost_state$state.name[tbl_mean_cost_state$state.name=="NuevoLeon"] <- "Nuevo Leon"
tbl_mean_cost_state$state.name[tbl_mean_cost_state$state.name=="Mexico City"] <- "Ciudad De Mexico"

tbl_mean_cost_state
```
### Mean cost per MWh Before Taxes and Fees
```{r}
tbl_mean_cost_state_notax <- dstrata %>% 
  group_by(state.name, is.summer1)  %>%
  summarize(mean_cost_mwh_notax = survey_mean(price.mWh.notax.usd, na.rm= TRUE, vartype = c("ci"))) 

tbl_mean_cost_state_notax <- na.omit(tbl_mean_cost_state_notax) 
colnames(tbl_mean_cost_state_notax)[which(names(tbl_mean_cost_state_notax) == "region.f")] <- "Region"
# renaming all the column names of data frame

#State names that are different in both dataframe
tbl_mean_cost_state_notax$state.name[tbl_mean_cost_state_notax$state.name=="NuevoLeon"] <- "Nuevo Leon"
tbl_mean_cost_state_notax$state.name[tbl_mean_cost_state_notax$state.name=="Mexico City"] <- "Ciudad De Mexico"

tbl_mean_cost_state_notax
```

```{r}
df_state_merge <- merge(x = tbl_mean_cost_state[ , c('state.name', 'is.summer1',
                                               'mean_cost_mwh')], 
                      y = df_clim_reg[ , c('state.name', 'state.abr', 
                                           'climatic.region', 'state_code')],
                      by = 'state.name', all.x=TRUE)

df_state_merge$dummy <- paste0(df_state_merge$state.name, df_state_merge$is.summer1)
tbl_mean_cost_state_notax$dummy <- paste0(tbl_mean_cost_state_notax$state.name, tbl_mean_cost_state_notax$is.summer1)

df_state_merge <- merge(x = df_state_merge[ , c('state.name', 'is.summer1', 'state.abr', 
                                           'climatic.region', 'state_code',
                                               'mean_cost_mwh','dummy')], 
                      y = tbl_mean_cost_state_notax[ , c('state.name', 'mean_cost_mwh_notax', 'dummy')],
                      by = c('dummy'), all.x=TRUE)

df_state_merge
```

```{r}
#load Mexico's map dataset from mxmaps project
data("mxstate.topoJSON")

tmpdir <- tempdir()
# have to use RJSONIO or else the topojson isn't valid
write(RJSONIO::toJSON(mxstate.topoJSON), file.path(tmpdir, "state.topojson"))
states <- topojson_read(file.path(tmpdir, "state.topojson"))

# convert to SpatialPolygonsDataFrame
states <- as_Spatial(states)
```


```{r}
#Make sure the coordinates of the labels are in the correct order
df_state_merge <- cbind(df_state_merge, 
                        data.frame(lon = coordinates(states[match(df_state_merge$state_code, states$id),])[,1],
                                   lat = coordinates(states[match(df_state_merge$state_code, states$id),])[,2]))
df_state_merge
```
**Obtain mean cost per MWh after taxes and fees**
```{r}
#df_state_summer <- df_state_merge[df_state_merge=='Yes']
df_state_merge$value <- df_state_merge$mean_cost_mwh
df_state_merge$region <- df_state_merge$state_code

df_state_summer <- df_state_merge[ which(df_state_merge$is.summer1 =='Yes'), ]
df_state_nosummer <- df_state_merge[ which(df_state_merge$is.summer1 =='No'), ]
df_state_summer
```

```{r}
plt_cost_summer <- mxstate_choropleth(df_state_summer, num_colors = 1) 
plt_cost_summer <- plt_cost_summer +
  scale_fill_gradientn("USD", labels = dollar, limits= c(25,100), colors = alpha(c(rev(brewer.pal(11, "Spectral"))), alpha = .7)) +
  labs(x = NULL, y = NULL,
       title = "Average per MWh Payment per Household After Fees and Taxes",
       subtitle = "Summer Months",
       caption = paste("Source: Own eleboration based on data from the",
                       "National Institute of Statistics and Geography (2018)", 
                       sep = "\n"))
plt_cost_summer <- plt_cost_summer +
  theme(plot.title = element_text(size = 16, face = "bold"),
        plot.caption = element_text(size = 9),
        legend.text = element_text(size = 10),
        legend.position = c(0.86, 0.75))

ggsave("./output/map_monthly_cost_mwh_summer.png", plot = plt_cost_summer, 
       dpi = 300, height = 6, width = 8, unit = 'in')

plt_cost_summer
```
```{r}
plt_cost_nosummer <- mxstate_choropleth(df_state_nosummer, num_colors = 1)

plt_cost_nosummer <- plt_cost_nosummer +
  scale_fill_gradientn("USD", labels = dollar, limits= c(25,100), 
                       colors = alpha(c(rev(brewer.pal(11, "Spectral"))), 
                                      alpha = .7)) +
  labs(x = NULL, y = NULL,
       title = "Average per MWh Payment per Household After Fees and Taxes",
       subtitle = "No Summer Months",
       caption = paste("Source: Own eleboration based on data from the",
                       "National Institute of Statistics and Geography (2018)", 
                       sep = "\n"))

plt_cost_nosummer <- plt_cost_nosummer +
  theme(plot.title = element_text(size = 16, face = "bold"),
        plot.caption = element_text(size = 9),
        legend.text = element_text(size = 10),
        legend.position = c(0.86, 0.75))

ggsave("./output/map_monthly_cost_mwh_nosummer.png", plot = plt_cost_summer, 
       dpi = 300, height = 6, width = 8, unit = 'in')

plt_cost_nosummer
```

**Obtain mean cost per MWh before taxes and fees**

```{r}
#df_state_summer <- df_state_merge[df_state_merge=='Yes']
df_state_merge$value <- df_state_merge$mean_cost_mwh_notax
df_state_merge$region <- df_state_merge$state_code

df_state_summer <- df_state_merge[ which(df_state_merge$is.summer1 =='Yes'), ]
df_state_nosummer <- df_state_merge[ which(df_state_merge$is.summer1 =='No'), ]
df_state_summer
```

```{r}
plt_cost_summer_notax <- mxstate_choropleth(df_state_summer, num_colors = 1) 
plt_cost_summer_notax <- plt_cost_summer_notax +
  scale_fill_gradientn("USD", labels = dollar, limits= c(25,100), colors = alpha(c(rev(brewer.pal(11, "Spectral"))), alpha = .7)) +
  labs(x = NULL, y = NULL,
       title = "Average per MWh Payment per Household Before Fees and Taxes",
       subtitle = "Summer Months",
       caption = paste("Source: Own eleboration based on data from the",
                       "National Institute of Statistics and Geography (2018)", 
                       sep = "\n"))
plt_cost_summer_notax <- plt_cost_summer_notax +
  theme(plot.title = element_text(size = 16, face = "bold"),
        plot.caption = element_text(size = 9),
        legend.text = element_text(size = 10),
        legend.position = c(0.86, 0.75))

ggsave("./output/map_monthly_cost_mwh_summer_notax.png", plot = plt_cost_summer_notax, 
       dpi = 300, height = 6, width = 8, unit = 'in')

plt_cost_summer_notax
```

```{r}
plt_cost_nosummer_notax <- mxstate_choropleth(df_state_nosummer, num_colors = 1) 
plt_cost_nosummer_notax <- plt_cost_nosummer_notax +
  scale_fill_gradientn("USD", labels = dollar, limits= c(25,100), colors = alpha(c(rev(brewer.pal(11, "Spectral"))), alpha = .7)) +
  labs(x = NULL, y = NULL,
       title = "Average per MWh Payment per Household Before Fees and Taxes",
       subtitle = "No Summer Months",
       caption = paste("Source: Own eleboration based on data from the",
                       "National Institute of Statistics and Geography (2018)", 
                       sep = "\n"))
plt_cost_nosummer_notax <- plt_cost_nosummer_notax +
  theme(plot.title = element_text(size = 16, face = "bold"),
        plot.caption = element_text(size = 9),
        legend.text = element_text(size = 10),
        legend.position = c(0.86, 0.75))

ggsave("./output/map_monthly_cost_mwh_nosummer_notax.png", plot = plt_cost_nosummer_notax, 
       dpi = 300, height = 6, width = 8, unit = 'in')

plt_cost_nosummer_notax
```
